+++ CHANGES.rst
@@ -2,6 +2,7 @@ Changes in version 1.1 (git)
 ----------------------------
 
 * Fixed Python 3 issue in the admin code
+* Fixed attempting to import south in Django 1.7/1.8 environments
 * Test ``polymorphic.__version__`` to determine the api of ``get_child_type_choice()``.